Move the forest layer panel to its own component
Moving the forest layer panel to its own component
and setting up the Map Visualizer Panel component
to support  a list of upcoming new sections.
