+++ src/client/pages/Geo/GeoMap/GeoMapMenuData/MapVisualizerPanel/ForestLayerSection/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ForestLayerSection'
